Stopgap to correctly orient the video, photo, and depth frames.
Video: Was oriented via .videoOrientation property in VideoCapture, now rotated in OverlayingTextureGenerater.
Depth: Continues to be oriented via .videoOrientation property in VideoCapture.
Photo: Was never correctly oriented, now rotated in OverlayingTextureGenerater.

Ideally, the orientations of all 3 would be configured the same way, but I'm unable to find a way to do so. A more feasible improvement would be to re-orient the video/photo frames in CameraTextureGenerater, rather than OverlayingTextureGenerater. The hack of using OverlayingTextureGenerater circumvents the new support for rotations added to MetalRenderingDevice.makeRenderVertexBuffer(), but that functionality can still be used in the future.