Ensure depths are announced with at most 0.1 precision.
Due to accumulated rounding errors that are inherent to floating-point arithmetic in Swift, we cannot guarantee that our rounding logic will yield values with the intended level of precision, regardless of whether we use Float or Double types. String formatting can guarantee the specified level of precision, so formatting the depth string prior to announcement ensures that the user only hears the correct level of precision.